Refactor home layout (#115)
* refactor: convert home screen to lazy column layout

* feat: enhance HomeScreen UI with sticky header and transaction grouping

This commit introduces several UI enhancements to the HomeScreen:

-   A sticky header is implemented, combining the account summary and toolbar.
-   Transactions are now grouped by date ("Today", "Yesterday", or specific date).
-   The overall visual design of the HomeScreen has been refined.
-   Updated `AccountSummaryUiModel` to use `BigDecimal` for balance.
-   The default transaction name now uses the category name if no specific name is provided.
-   The transaction item now displays `toAccountName`, `merchantName`, or `fromAccountName` before falling back to `name`.

+++ app/src/main/java/dev/pandesal/sbp/presentation/model/AccountSummaryUiModel.kt
@@ -1,8 +1,10 @@
 package dev.pandesal.sbp.presentation.model
 
+import java.math.BigDecimal
+
 data class AccountSummaryUiModel(
     val name: String,
-    val balance: Double,
+    val balance: BigDecimal,
     val isSpendingWallet: Boolean,
     val isFundingWallet: Boolean,
     val currency: String
